+++ mobile/android/android-components/components/feature/toolbar/src/main/java/mozilla/components/feature/toolbar/ToolbarFeature.kt
@@ -4,9 +4,11 @@
 
 package mozilla.components.feature.toolbar
 
+import android.support.annotation.ColorInt
 import mozilla.components.concept.toolbar.Toolbar
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.feature.session.SessionUseCases
+import mozilla.components.lib.publicsuffixlist.PublicSuffixList
 import mozilla.components.support.base.feature.BackHandler
 import mozilla.components.support.base.feature.LifecycleAwareFeature
 
@@ -24,9 +26,10 @@ class ToolbarFeature(
     sessionManager: SessionManager,
     loadUrlUseCase: SessionUseCases.LoadUrlUseCase,
     searchUseCase: SearchUseCase? = null,
-    sessionId: String? = null
+    sessionId: String? = null,
+    urlRenderConfiguration: UrlRenderConfiguration? = null
 ) : LifecycleAwareFeature, BackHandler {
-    private val presenter = ToolbarPresenter(toolbar, sessionManager, sessionId)
+    private val presenter = ToolbarPresenter(toolbar, sessionManager, sessionId, urlRenderConfiguration)
     private val interactor = ToolbarInteractor(toolbar, loadUrlUseCase, searchUseCase)
 
     /**
@@ -52,4 +55,18 @@ class ToolbarFeature(
     override fun stop() {
         presenter.stop()
     }
+
+    /**
+     * Configuration that controls how URLs are rendered.
+     *
+     * @property publicSuffixList A shared/global [PublicSuffixList] object required to extract certain domain parts.
+     * @property registrableDomainColor Text color that should be used for the registrable domain of the URL (see
+     * [PublicSuffixList.getPublicSuffixPlusOne] for an explanation of "registrable domain".
+     * @property urlColor Optional text color used for the URL.
+     */
+    data class UrlRenderConfiguration(
+        internal val publicSuffixList: PublicSuffixList,
+        @ColorInt internal val registrableDomainColor: Int,
+        @ColorInt internal val urlColor: Int? = null
+    )
 }