+++ mobile/android/android-components/components/feature/toolbar/src/main/java/mozilla/components/feature/toolbar/ToolbarPresenter.kt
@@ -4,20 +4,27 @@
 
 package mozilla.components.feature.toolbar
 
+import android.support.annotation.VisibleForTesting
 import mozilla.components.browser.session.SelectionAwareSessionObserver
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.concept.toolbar.Toolbar
+import mozilla.components.feature.toolbar.internal.URLRenderer
 
 /**
  * Presenter implementation for a toolbar implementation in order to update the toolbar whenever
  * the state of the selected session changes.
  */
+@Suppress("TooManyFunctions")
 class ToolbarPresenter(
     private val toolbar: Toolbar,
     private val sessionManager: SessionManager,
-    private val sessionId: String? = null
+    private val sessionId: String? = null,
+    urlRenderConfiguration: ToolbarFeature.UrlRenderConfiguration? = null
 ) : SelectionAwareSessionObserver(sessionManager) {
+
+    @VisibleForTesting
+    internal var renderer = URLRenderer(toolbar, urlRenderConfiguration)
 
     /**
      * Start presenter: Display data in toolbar.
@@ -25,6 +32,14 @@ class ToolbarPresenter(
     fun start() {
         observeIdOrSelected(sessionId)
         initializeView()
+
+        renderer.start()
+    }
+
+    override fun stop() {
+        super.stop()
+
+        renderer.stop()
     }
 
     /**
@@ -49,13 +64,13 @@ class ToolbarPresenter(
         val session = sessionId?.let { sessionManager.findSessionById(sessionId) }
             ?: sessionManager.selectedSession
 
-        toolbar.url = session?.url ?: ""
+        renderer.post(session?.url ?: "")
         toolbar.displayProgress(session?.progress ?: 0)
         updateToolbarSecurity(session?.securityInfo ?: Session.SecurityInfo())
     }
 
     override fun onUrlChanged(session: Session, url: String) {
-        toolbar.url = url
+        renderer.post(url)
         toolbar.setSearchTerms(session.searchTerms)
     }