+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,19 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
 
+* **feature-toolbar**
+  * Added ability to color parts of the domain (e.g. [registrable domain](https://url.spec.whatwg.org/#host-registrable-domain)) by providing a `UrlRenderConfiguration`:
+
+  ```kotlin
+  ToolbarFeature(
+    // ...
+    ToolbarFeature.UrlRenderConfiguration(
+        publicSuffixList, // Use a shared global instance
+        registrableDomainColor = 0xFFFF0000.toInt(),
+        urlColor = 0xFF00FF00.toInt()
+    )
+  ```
+
 # 0.44.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.43.0...v0.44.0)